Remove nightly feature
See #19 for rationale. README has been heavily expanded to reflect
changes.
