+++ README.md
@@ -21,7 +21,7 @@ Mostly attributes are removed.
 
 ## Usage
 
-This paragraph describes how to use this crate correctly to ensure
+This section describes how to use this crate correctly to ensure
 compatibility and interoperability on both stable and nightly channels.
 
 If you are writing a library that interacts with allocators API, you can
@@ -29,12 +29,62 @@ add this crate as a dependency and use the types and traits from this
 crate instead of the ones in `core` or `alloc`.
 This will allow your library to compile on stable and beta channels.
 
-Your library *MAY* provide a feature that will enable "allocator-api2/nightly".
-When this feature is enabled, your library *MUST* enable
-unstable `#![feature(allocator_api)]` or it may not compile.
-If feature is not provided, your library may not be compatible with the
-rest of the users and cause compilation errors on nightly channel
-when some other crate enables "allocator-api2/nightly" feature.
+If you wish to be able to use Rust's unstable [`allocator_api`] in place of the
+API provided by this crate when your crate is built on the nightly channel,
+you can do so by adding a `nightly` feature to your crate and taking the
+following steps.
+
+Add the following lines to the crate root:
+
+```rust
+#![cfg_attr(feature = "nightly", feature(allocator_api))]
+#[cfg(feature = "nightly")]
+extern crate alloc;
+```
+
+The `extern crate alloc` statement can be replaced with `extern crate core` if
+only items from `core` are required, and if the desired crate is already being
+pulled in otherwise, it need not be repeated with the
+`#[cfg(feature = "nightly")]` flag.
+
+Next, wherever `allocator-api2` is used, lock the import behind a
+`#[cfg(not(feature = "nightly"))]` statement and add an import of the equivalent
+unstable [`allocator_api`] item, locked behind a `#[cfg(feature = "nightly")]`
+statement. For example:
+
+```rust
+#[cfg(not(feature = "nightly"))]
+use allocator_api2::alloc::Allocator;
+#[cfg(feature = "nightly")]
+use alloc::alloc::Allocator;
+```
+
+Alternatively, one could add similar statements to a module, export them with
+`pub use` and import the required items from the module, in order to avoid
+cluttering the rest of the code with `#[cfg]` statements.
+
+With this setup, the `nightly` feature will, when enabled, demand that the crate
+be built on the nightly channel and use the unstable API instead of the one
+provided by `allocator-api2`. Note that this is not necessary to build a crate
+depending on `allocator-api2` on the nightly channel: Without this setup, the
+crate will simply continue to use the API provided by `allocator-api2` even when
+built on the nightly channel.
+
+### Note about older `allocator-api2` versions
+
+In `allocator-api2` versions `0.2` and below, a `nightly` feature was provided
+which re-exported the unstable [`allocator_api`] in replacement of the normal
+API provided by this crate. This was found to be problematic, as any crate
+utilizing `allocator-api2` without providing a feature to enable
+`#![feature(allocator_api)]` would be unable to compile if another crate in the
+tree enabled the `allocator-api2/nightly` feature. And even when such a feature
+was provided, other crates transitively depending on such crates may have had to
+pull in their transitive dependencies explicitly in order to enable it. For
+these reasons, the `nightly` feature of this crate has been removed. Users of
+any of these older versions are urged to update to a newer version of
+`allocator-api2` to ensure compatibility with other crates. The instructions
+above have been provided as a substitute for those wishing to be able to switch
+between using the API provided by this crate and the unstable [`allocator_api`].
 
 # Minimal Supported Rust Version (MSRV)
