--- src/nightly.rs
@@ -1,5 +1,0 @@
-#[cfg(not(feature = "alloc"))]
-pub use core::alloc;
-
-#[cfg(feature = "alloc")]
-pub use alloc_crate::{alloc, boxed, vec, collections};
