+++ src/vec/set_len_on_drop.rs
